Moved handleViews function to PostsContext
