+++ README.md
@@ -147,4 +147,5 @@ Below is an explanation of all included `npm` commands in the root `package.json
 5. I ran into a bug where I would click on a category and it would load the posts from that category, but then if I clicked on the Home link, it would only show the posts of the categories that I just navigated from. The problem was my state never got reset so it's still referencing the old category.
 6. My middleware auth had an issue where my user token was not being saved to localstorage even though the token was being returned from the api on the server side. The issue was my React Router in my `App` component on the client side was not wrapped with my UserProvider function defined in the context, UserContext component.
 7. I was making two different API calls in my `App` component to conditionally load posts by all or by category. I ran into an issue where I need to load cards with the posts table, but also join the users table to show the usernames and their respective posts. So I created a type UserPost and then used a union type for my state which would hold the posts retrieved from the API calls.
-8. Discovered an issue where all the posts were being rendered in `App`, but since App mounts once and doesn't unmount, a create post action doesn't automatically fetch new posts and update the `App` component. So the post form sends the user to the home page but the new post isn't seen until refreshing the page. I ended up refactoring my code to use context and create a PostsContext provider which held the fetch posts functions and state. Then in my components that needed posts, I destructured my posts object from usePosts hook. This also broke my `NavBar` component since it was using an event handler function in App which I moved the fetch posts functions away from `App` to my usePosts hook.
+8. Discovered an issue where all the posts were being rendered in `App`, but since App mounts once and doesn't unmount, a create post action doesn't automatically fetch new posts and update the `App` component. So the post form sends the user to the home page but the new post isn't seen until refreshing the page. I ended up refactoring my code to use context and create a PostsContext provider which held the fetch posts functions and state. Then in my components that needed posts, I destructured my posts from the context provider. This also broke my `NavBar` component since it was using an event handler function in App which I moved the fetch posts functions away from `App` to my usePosts hook.
+9. Issue where creating a post would create the post in the database but the new post wouldn't get fetched on the client side. I had to redo my database schema so that the foreign keys on tables had the exact name for primary keys on other tables. For example users.userId -> post.userId.
