+++ client/src/pages/Home.tsx
@@ -2,41 +2,15 @@ import { Card } from '../components/Card';
 import { Category } from '../lib/data';
 import { NavBar } from '../components/NavBar';
 import { usePosts } from '../components/usePosts';
-import { useState } from 'react';
-import { useUser } from '../components/useUser';
-import { Posts } from '../components/PostsContext';
 
 type Props = {
   isMobile: boolean | null;
   categories: Category[];
+  error: unknown;
 };
 
-export function Home({ isMobile, categories }: Props) {
-  const [error, setError] = useState<unknown>();
-  const { posts } = usePosts();
-  const { token } = useUser();
-
-  async function handleViews(post: Posts) {
-    try {
-      const updatedPost = {
-        ...post,
-        views: post.views++,
-      };
-      const req = {
-        method: 'PUT',
-        headers: {
-          Authorization: `Bearer ${token}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(updatedPost),
-      };
-      const res = await fetch(`/api/posts/${post.postId}`, req);
-      if (!res.ok) throw new Error(`fetch Error: ${res.status}`);
-    } catch (error) {
-      setError(error);
-    }
-  }
-
+export function Home({ isMobile, categories, error }: Props) {
+  const { posts, handleViews } = usePosts();
   if (error || !posts) {
     return (
       <div>
